minor refactor splitting up klab-build and klab-debug dependencies
