+++ lib/behavior.js
@@ -1,6 +1,6 @@
 const C = require("./constraints.js")
 const S = require("./state.js")
-const kast = require("./kast.js");
+const kast = require("./util/kast.js");
 
 const genBehaviour = (state) => {
   const FILTER_OOG = state.config.FILTER_OOG || false;
