+++ lib/compile.js
@@ -3,9 +3,9 @@ const path = require("path");
 const {
   testPath,
   read
-} = require("../lib/util.js");
+} = require("../lib/util/util.js");
 const pure = require("../lib/pure.js");
-const kast = require("./kast.js");
+const kast = require("./util/kast.js");
 
 // TODO - throw error if TMPDIR is not set
 const KLAB_WD_PATH = path.join(process.env.TMPDIR, "klab");
