+++ lib/constraints.js
@@ -1,4 +1,4 @@
-const kast = require("./kast.js")
+const kast = require("./util/kast.js")
 
 const clean = cs => cs
   .map(s => s.replace(/ ==K true/g,"").trim())
