+++ lib/driver/dumpDriver.js
@@ -4,7 +4,7 @@ const path = require("path");
 const {
   testPath,
   ensureDir
-} = require("../util.js");
+} = require("../util/util.js");
 
 const KLAB_OUT = process.env.KLAB_OUT || "out";
