+++ lib/driver/kDriver.js
@@ -7,7 +7,7 @@ const moment = require("moment");
 const clc    = require("cli-color");
 const {
   testPath
-} = require("../util.js");
+} = require("../util/util.js");
 
 const KLAB_WD_PATH = path.join(process.env.TMPDIR, "klab");
 if(!testPath(KLAB_WD_PATH)) fs.mkdirSync(KLAB_WD_PATH);
