+++ lib/driver/makeDriver.js
@@ -7,7 +7,7 @@ const { spawnSync } = require('child_process');
 const {
   testPath,
   ensureDir
-} = require("../util.js");
+} = require("../util/util.js");
 
 const KLAB_WD_PATH = path.join(process.env.TMPDIR, "klab");
 const KLAB_EVMS_PATH = process.env.KLAB_EVMS_PATH;
