+++ lib/evmv.js
@@ -1,4 +1,4 @@
-const kast = require("./kast.js");
+const kast = require("./util/kast.js");
 const {
   buildDisplayInfo
 } = require("../lib/compile.js");
@@ -45,6 +45,7 @@ module.exports = function (state) {
   } = buildDisplayInfo(k, state.config);
 
 
+  //Extract bin_runtime from program cell
   // console.log(bin_runtime);
   let spec_o = state.config.rule_meta
   let contract_o = state.config.implementations[spec_o.v2n[call_id]];
