+++ lib/modules.js
@@ -1,11 +1,9 @@
 const clc = require('cli-color');
-// const { formatStep } = require("./formatter.js");
 const { formatMemory } = require("./unwrapper.js");
 const C = require("./constraints.js");
 const S = require("./state.js");
 const Rule = require("./rule.js");
-const kast = require("./kast.js");
-// const Constraint = require("./constraints.js")
+const kast = require("./util/kast.js");
 const format = require("./format.js")
 const { genBehaviour } = require("./behavior.js");
 const evmv = require("./evmv.js");
@@ -69,7 +67,11 @@ const view = {
       let spec_o = state.config.rule_meta;
       let contract_o = state.config.implementations[spec_o.v2n[call_id]];
       let contract = state.config.contracts[contract_o.name];
-      let src = getCodeStringFromPc(state.config.srcs, contract, parseInt(pc), true);
+      let src;
+      //console.log(contract.srcmapArr === undefined)
+      src = (contract.srcmapmapArr === undefined) ?
+        getCodeStringFromPc(state.config.srcs, contract, parseInt(pc), true)
+        : 'source not available';
       return src;
     }
   },
