+++ lib/srchandler.js
@@ -2,14 +2,6 @@ const fs = require("fs");
 const clc = require('cli-color');
 const tw = process.stdout.columns; // total width
 
-// EVM Instructions
-const evm_i = require("./evm.json");
-
-
-// HELPER
-const hex = i => (s = i.toString(16)).length == 1 ? "0" + s : s;
-
-
 const genSrcmapArr = srcmap => srcmap
   .split(";")
   .map(l => l.split(":"))
@@ -29,37 +21,6 @@ const genSrcmapArr = srcmap => srcmap
   )
   .reverse()
   .slice(1)
-
-// Instructions
-const get_pc_to_inst_map = bin_string => {
-  const code_arr = []
-  for(let i=0; i<bin_string.length; i+=2) {
-    code_arr.push(parseInt(bin_string.slice(i, i+2), 16));
-  }
-  const pc_to_inst_map = [];
-  const instructions = [];
-  let i = 0;
-  while(i < code_arr.length) {
-    let i_ = i;
-    let index = hex(code_arr[i]);
-    let instruction_string = evm_i[index] || `UNKNOWN (${index})`;
-
-    if(code_arr[i] >= 96 && code_arr[i] <= 127) {
-      let length = code_arr[i] - 95;
-      instruction_string += " " + code_arr.slice(i + 1, i + 1 + length).map(s => hex(s)).join("");
-      i += length;
-    }
-    i++;
-    instructions.push(instruction_string);
-    for(let j=i_; j < i; j++) {
-      pc_to_inst_map.push(instructions.length - 1);
-    }
-  }
-  return {
-    pc_to_inst_map,
-    instructions
-  };
-}
 
 // EXPORT
 // This mess deals with displaying the highlighted compact
@@ -150,61 +111,7 @@ const getCodeStringFromPc = (srcs, contract, pc, trim) => {
   return str;
 }
 
-let getNodeSrcAst = (ast, name) => {
-    return ast.map(node => {
-        if (node.attributes && node.attributes.name) {
-            if (node.attributes.name == name) {
-                return node.src;
-            }
-        }
-        if (node.children) {
-            return getNodeSrcAst(node.children, name).join('');
-        }
-    })
-}
-
-let getContractAst = (ast, contractName) => {
-    return ast.map(node => {
-        if (node.attributes && node.attributes.name) {
-            if (node.attributes.name == contractName) {
-                return node;
-            }
-        }
-        if (node.children) {
-            return getContractAst(node.children, contractName).join('');
-        }
-    })
-
-}
-
-let functionNameToPcRange = (name, srcmapArr, ast, bin_runtime, inst_to_pc) => {
-    let pcList = get_pc_to_inst_map(bin_runtime).pc_to_inst_map;
-    let srcFromAst = getNodeSrcAst([ast], name)[0];
-    let s_l_f = srcmapArr.map(e => [e.s, e.l, e.f].join(':'));
-    let start_index = s_l_f.indexOf(srcFromAst)
-    //Find the corresponding jump 'out'
-    //each jump 'in' increases depth by 1.
-    let depth = 1;
-    let j = srcmapArr.slice(start_index).map(e => e.j)
-    let i = 0
-    while (depth > 0 && i < j.length) {
-        if (j[i] == 'i') {
-            depth++;
-        } else if (j[i] == 'o') {
-            depth--;
-        }
-        i++;
-    }
-    let end_index = Number(start_index) + i -1 ;
-    let pc_start = inst_to_pc[start_index]
-    let pc_end = inst_to_pc[end_index]
-    return [pc_start, pc_end]
-}
-
 module.exports = {
   getCodeStringFromPc,
-  get_pc_to_inst_map,
-  genSrcmapArr,
-  functionNameToPcRange,
-  getContractAst  
+  genSrcmapArr
 };
