+++ lib/state.js
@@ -1,5 +1,5 @@
 const Constraints = require("./constraints.js");
-const {format} = require("./kast.js");
+const {format} = require("./util/kast.js");
 
 const term_id = state => {
   return state.path[state.path.length - 1].step.to.split("_")[0];
