+++ lib/util/kast.js
