+++ lib/util/locale.js
