+++ lib/util/util.js
@@ -3,6 +3,8 @@ const fs     = require("fs");
 const path   = require("path");
 const locale = require("./locale.js");
 const keccak = require("keccak");
+// EVM Instructions
+const evm_i = require("../../resources/evm.json");
 const sha3 = function (str) {
   return keccak('keccak256')
     .update(str)
@@ -81,6 +83,54 @@ const toK = str => str
   .replace(/bool\(/g, "bool2Word(")
 
 
+// HELPER
+const hex = i => (s = i.toString(16)).length == 1 ? "0" + s : s;
+
+
+// Instructions
+const get_pc_to_inst_map = bin_string => {
+  const code_arr = []
+  for(let i=0; i<bin_string.length; i+=2) {
+    code_arr.push(parseInt(bin_string.slice(i, i+2), 16));
+  }
+  const pc_to_inst_map = [];
+  const instructions = [];
+  let i = 0;
+  while(i < code_arr.length) {
+    let i_ = i;
+    let index = hex(code_arr[i]);
+    let instruction_string = evm_i[index] || `UNKNOWN (${index})`;
+
+    if(code_arr[i] >= 96 && code_arr[i] <= 127) {
+      let length = code_arr[i] - 95;
+      instruction_string += " " + code_arr.slice(i + 1, i + 1 + length).map(s => hex(s)).join("");
+      i += length;
+    }
+    i++;
+    instructions.push(instruction_string);
+    for(let j=i_; j < i; j++) {
+      pc_to_inst_map.push(instructions.length - 1);
+    }
+  }
+  return {
+    pc_to_inst_map,
+    instructions
+  };
+}
+
+let getNodeSrcAst = (ast, name) => {
+    return ast.map(node => {
+        if (node.attributes && node.attributes.name) {
+            if (node.attributes.name == name) {
+                return node.src;
+            }
+        }
+        if (node.children) {
+            return getNodeSrcAst(node.children, name).join('');
+        }
+    })
+}
+
 module.exports = {
   ensureDirs,
   ensureDir,
@@ -92,5 +142,7 @@ module.exports = {
   warn,
   mapInterface,
   toK,
-  getId
+  getId,
+  get_pc_to_inst_map,
+  getNodeSrcAst
 };
