--- libexec/klab-evm
@@ -1,131 +1,0 @@
-#!/usr/bin/env node
-
-const xml2json = require("xml2json");
-
-const fs = require("fs");
-const _ = require("lodash");
-const jsondiffpatch = require("jsondiffpatch").create({
-   textDiff: {
-        // default 60, minimum string length (left and right sides) to use text diff algorythm: google-diff-match-patch
-        minLength: 10
-    }
-});
-const clc = require('cli-color');
-var clear = require("cli-clear");
-const testPath = path => {
-  try {
-    fs.accessSync(path, fs.constants.F_OK);
-    return true;
-  } catch (e) {
-    return false;
-  }
-}
-
-const { formatStep } = require("./../lib/formatter.js");
-
-var stdin = process.stdin;
-stdin.setRawMode( true );
-stdin.resume();
-stdin.setEncoding( 'utf8' );
-const tw = process.stdout.columns; // total width
-
-const config = JSON.parse(fs.readFileSync("./config.json"));
-if (testPath(config.solJson)) {
-    config.solJson = JSON.parse(fs.readFileSync(config.solJson));
-    contract_path = Object
-      .keys(config.solJson.contracts)
-      .find(name => (new RegExp(config.name + "$")).test(name))
-    // contract_path = config
-    //   .source
-    //   .replace('/','')
-    //   .replace('.','')
-    // + config
-    //   .source
-    //   .replace('./',':')
-    //   .replace('.sol','');
-    contract_json = config.solJson.contracts[contract_path];
-    config["bin_runtime"] = contract_json["bin-runtime"].toString();
-    config["sourcemap"]   = contract_json["srcmap-runtime"].toString();
-} else {
-    config["bin_runtime"] = fs.readFileSync(config["bin_runtime"]).toString()
-    config["sourcemap"]   = fs.readFileSync(config["sourcemap"]).toString()
-}
-
-
-
-// READ the spec files
-[
-  "spec",
-  "lemmas",
-  "source",
-].forEach(filetype => {
-  let content = fs.readFileSync(config[filetype]).toString()
-  config[filetype] = content;
-})
-
-const {
-  getCodeStringFromPc,
-  get_pc_to_inst_map,
-  genSrcmapArr
-} = require("../lib/srchandler.js");
-const {
-  instructions,
-  pc_to_inst_map
-} = get_pc_to_inst_map(config.bin_runtime)
-const srcmapArr = genSrcmapArr(config.sourcemap);
-const inst_to_pc = instructions.map((inst, i) => pc_to_inst_map.indexOf(i));
-
-const state = {
-  inst: 0
-};
-
-const r = s => " ".repeat(4 - s.length) + s
-const hex = n => n.toString(16).length % 2 == 0 ? n.toString(16) : "0" + n.toString(16);
-
-const draw = () => {
-  let pc = inst_to_pc[state.inst];
-  let from = Math.max(0, state.inst - 8);
-  let to   = Math.min(instructions.length, from + 17);
-  let str = instructions
-    .slice(from, to)
-    .map((s, i) => (from + i === state.inst ? "> " : "  ") + r(hex(inst_to_pc[from + i])) + "  " + r(hex(from + i)) + "  " + (s.length > 20 ? s.slice(0, 20) + "..." : s))
-    .join("\n")
-  str = str +"\n\n"+ getCodeStringFromPc(config, pc, true, pc_to_inst_map, srcmapArr)
-
-  let status_line = ` PC: ${pc} (0x${hex(pc)})`;
-  status_line += " ".repeat(tw - status_line.length);
-  status_line = clc.bgXterm(236)(status_line);
-
-  let header = ("    PC    ID  Opcode");
-  console.log(status_line + `\n\n${header}\n` + str);
-}
-
-stdin.on( 'data', function (key, e) {
-  clear();
-  if(key === "q" || key === "\u0003") {
-    process.exit();
-  } else if(key === "\u001B\u005B\u0041" // UP
-    || key === "k"
-  ) {
-    if(state.inst > 0) {
-      state.inst = state.inst - 1;
-    }
-  } else if(key === "\u001B\u005B\u0042"
-    || key === "j"
-  ) {
-    if(state.inst < instructions.length) {
-      state.inst = state.inst + 1;
-    }
-  } else if(key === "\u0004") {
-    state.inst = Math.min(instructions.length - 1, state.inst + 8)
-  } else if(key === "\u0015") {
-    state.inst = Math.max(0, state.inst - 8)
-  } else {
-    console.log("Unknown key:" + key.split("").map(s => s.charCodeAt(0).toString(16)));
-  }
-  draw();
-})
-
-clear();
-draw();
-
